+++ ui/src/alerting/components/ChecksColumn.tsx
@@ -7,14 +7,17 @@ import {connect} from 'react-redux'
 import {viewableLabels} from 'src/labels/selectors'
 
 // Components
+import {Button, IconFont, ComponentColor} from '@influxdata/clockface'
 import CheckCards from 'src/alerting/components/CheckCards'
 import AlertsColumnHeader from 'src/alerting/components/AlertsColumn'
 
 // Types
-import {Check, AppState} from 'src/types'
+import {Check, NotificationRuleDraft, AppState} from 'src/types'
 
 interface StateProps {
   checks: Check[]
+  rules: NotificationRuleDraft[]
+  endpoints: AppState['endpoints']['list']
 }
 
 type Props = StateProps & WithRouterProps
@@ -23,13 +26,46 @@ const ChecksColumn: FunctionComponent<Props> = ({
   checks,
   router,
   params: {orgID},
+  rules,
+  endpoints,
 }) => {
   const handleClick = () => {
     router.push(`/orgs/${orgID}/alerting/checks/new`)
   }
+
+  const tooltipContents = (
+    <>
+      A <strong>Check</strong> is a periodic query that the system
+      <br />
+      performs against your time series data
+      <br />
+      that will generate a status
+    </>
+  )
+
+  const noAlertingResourcesExist =
+    !checks.length && !rules.length && !endpoints.length
+
+  const createButton = (
+    <Button
+      color={ComponentColor.Primary}
+      text="Create"
+      onClick={handleClick}
+      testID="create-check"
+      icon={IconFont.Plus}
+    />
+  )
+
   return (
-    <AlertsColumnHeader title="Checks" onCreate={handleClick}>
-      <CheckCards checks={checks} />
+    <AlertsColumnHeader
+      title="Checks"
+      createButton={createButton}
+      questionMarkTooltipContents={tooltipContents}
+    >
+      <CheckCards
+        checks={checks}
+        showFirstTimeWidget={noAlertingResourcesExist}
+      />
     </AlertsColumnHeader>
   )
 }
@@ -38,9 +74,16 @@ const mstp = (state: AppState) => {
   const {
     checks: {list: checks},
     labels: {list: labels},
+    rules: {list: rules},
+    endpoints,
   } = state
 
-  return {checks, labels: viewableLabels(labels)}
+  return {
+    checks,
+    labels: viewableLabels(labels),
+    rules,
+    endpoints: endpoints.list,
+  }
 }
 
 export default connect<StateProps, {}, {}>(