+++ ui/src/alerting/components/EndpointsColumn.tsx
@@ -4,6 +4,7 @@ import {connect} from 'react-redux'
 import {withRouter, WithRouterProps} from 'react-router'
 
 // Components
+import {Button, IconFont, ComponentColor} from '@influxdata/clockface'
 import EndpointCards from 'src/alerting/components/endpoints/EndpointCards'
 import AlertsColumn from 'src/alerting/components/AlertsColumn'
 import {AppState} from 'src/types'
@@ -20,11 +21,31 @@ const EndpointsColumn: FC<Props> = ({router, params, endpoints}) => {
     router.push(newRuleRoute)
   }
 
+  const tooltipContents = (
+    <>
+      A <strong>Notification Endpoint</strong> stores the information to connect
+      <br />
+      to a third party service that can receive notifications
+      <br />
+      like Slack, PagerDuty, or an HTTP server
+    </>
+  )
+
+  const createButton = (
+    <Button
+      color={ComponentColor.Secondary}
+      text="Create"
+      onClick={handleOpenOverlay}
+      testID="create-endpoint"
+      icon={IconFont.Plus}
+    />
+  )
+
   return (
     <AlertsColumn
       title="Notification Endpoints"
-      testID="create-endpoint"
-      onCreate={handleOpenOverlay}
+      createButton={createButton}
+      questionMarkTooltipContents={tooltipContents}
     >
       <EndpointCards endpoints={endpoints} />
     </AlertsColumn>