+++ ui/src/alerting/components/notifications/RulesColumn.tsx
@@ -7,11 +7,18 @@ import {withRouter, WithRouterProps} from 'react-router'
 import {NotificationRuleDraft, AppState} from 'src/types'
 
 // Components
+import {
+  Button,
+  IconFont,
+  ComponentColor,
+  ComponentStatus,
+} from '@influxdata/clockface'
 import NotificationRuleCards from 'src/alerting/components/notifications/RuleCards'
 import AlertsColumn from 'src/alerting/components/AlertsColumn'
 
 interface StateProps {
   rules: NotificationRuleDraft[]
+  endpoints: AppState['endpoints']['list']
 }
 
 type Props = StateProps & WithRouterProps
@@ -20,17 +27,48 @@ const NotificationRulesColumn: FunctionComponent<Props> = ({
   rules,
   router,
   params,
+  endpoints,
 }) => {
   const handleOpenOverlay = () => {
     const newRuleRoute = `/orgs/${params.orgID}/alerting/rules/new`
     router.push(newRuleRoute)
   }
 
+  const tooltipContents = (
+    <>
+      A <strong>Notification Rule</strong> will query statuses
+      <br />
+      written by <strong>Checks</strong> to determine if a notification
+      <br />
+      should be sent to a <strong>Notification Endpoint</strong>
+    </>
+  )
+
+  const buttonStatus = !!endpoints.length
+    ? ComponentStatus.Default
+    : ComponentStatus.Disabled
+
+  const buttonTitleText = !!endpoints.length
+    ? 'Create a Notification Rule'
+    : 'You need at least 1 Notifcation Endpoint to create a Notification Rule'
+
+  const createButton = (
+    <Button
+      color={ComponentColor.Secondary}
+      text="Create"
+      onClick={handleOpenOverlay}
+      testID="create-rule"
+      icon={IconFont.Plus}
+      status={buttonStatus}
+      titleText={buttonTitleText}
+    />
+  )
+
   return (
     <AlertsColumn
       title="Notification Rules"
-      testID="create-rule"
-      onCreate={handleOpenOverlay}
+      createButton={createButton}
+      questionMarkTooltipContents={tooltipContents}
     >
       <NotificationRuleCards rules={rules} />
     </AlertsColumn>
@@ -40,9 +78,10 @@ const NotificationRulesColumn: FunctionComponent<Props> = ({
 const mstp = (state: AppState) => {
   const {
     rules: {list: rules},
+    endpoints,
   } = state
 
-  return {rules}
+  return {rules, endpoints: endpoints.list}
 }
 
 export default connect<StateProps, {}, {}>(